+++ Framework/FlamingoTests/OptionalProtocolTestCase.swift
@@ -7,7 +7,7 @@
 //
 
 import XCTest
-@testable import Flamingo
+import Flamingo
 
 class OptionalProtocolTestCase: XCTestCase {
     private struct Consts {
