+++ Framework/FlamingoTests/ValidationTests.swift
@@ -7,7 +7,7 @@
 //
 
 import XCTest
-@testable import Flamingo
+import Flamingo
 
 class RealFailedTestRequest: NetworkRequest {
