+++ Source/StubsSessionFactory.swift
@@ -8,6 +8,51 @@
 
 import Foundation
 
+public struct RequestStubMap: Decodable {
+    public let url: URL
+    public let method: HTTPMethod
+    public let params: [String: Any]?
+    public let responseStub: ResponseStub
+    public var requestStub: RequestStub {
+        return RequestStub(url: url, method: method, params: params)
+    }
+
+    private enum CodingKeys: String, CodingKey {
+        case url
+        case method
+        case params
+        case responseStub = "stub"
+    }
+
+    public init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        url = try container.decode(.url)
+        method = try container.decode(.method)
+        if container.contains(.params),
+            let paramsAsString: String = try container.decode(.params),
+            let paramsAsData = paramsAsString.data(using: .utf8) {
+            params = try JSONSerialization.jsonObject(with: paramsAsData, options: .allowFragments) as? [String: Any]
+        } else {
+            params = nil
+        }
+        responseStub = try container.decode(.responseStub)
+    }
+
+    public init(url: URL, method: HTTPMethod, params: [String: Any]?, responseStub: ResponseStub) {
+        self.url = url
+        self.method = method
+        self.params = params
+        self.responseStub = responseStub
+    }
+
+    public init(request: RequestStub, responseStub: ResponseStub) {
+        self.url = request.url
+        self.method = request.method
+        self.params = request.params
+        self.responseStub = responseStub
+    }
+}
+
 public struct StubFile: Decodable {
     public let version: String?
     public let stubs: [RequestStubMap]
@@ -31,24 +76,24 @@ public struct StubFile: Decodable {
 }
 
 public class StubsSessionFactory {
-    public static func session() -> StubDefaultSession {
-        return StubDefaultSession()
+    public static func session() -> StubsDefaultSession {
+        return StubsDefaultSession()
     }
 
-    public static func session(_ key: RequestStub, stub: ResponseStub) -> StubDefaultSession {
+    public static func session(_ key: RequestStub, stub: ResponseStub) -> StubsDefaultSession {
         let session = self.session()
         session.add(key, stub: stub)
         return session
     }
 
-    public static func session(with stubs: [RequestStubMap]) -> StubDefaultSession {
+    public static func session(with stubs: [RequestStubMap]) -> StubsDefaultSession {
         let session = self.session()
-        session.add(stubs: stubs)
+        session.add(stubsArray: stubs)
 
         return session
     }
 
-    public static func session(fromFile path: String, decoder: JSONDecoder = JSONDecoder()) throws -> StubDefaultSession {
+    public static func session(fromFile path: String, decoder: JSONDecoder = JSONDecoder()) throws -> StubsDefaultSession {
 
         let stubFile = try StubFile(fromFile: path, decoder: decoder)
